added linkify support to the availability texts
Yes I know I should be using Linkify but the destination the links go to
arent what the text is in the textview. The links go to the sign up page
for each service as @b3ngr33ni3r has told me. It ain't pretty but it
kinda sorta works.